+++ src/main/java/org/sopt/domain/auth/application/service/AuthService.java
@@ -5,14 +5,14 @@ import org.sopt.domain.auth.application.dto.AuthTokenResult;
 import org.sopt.domain.auth.domain.exception.AuthErrorCode;
 import org.sopt.domain.auth.domain.model.RefreshToken;
 import org.sopt.domain.auth.domain.repository.RefreshTokenRepository;
+import org.sopt.domain.auth.infrastructure.RefreshTokenHasher;
 import org.sopt.domain.user.domain.exception.UserErrorCode;
 import org.sopt.domain.user.domain.model.User;
 import org.sopt.domain.user.domain.repository.UserRepository;
 import org.sopt.global.exception.BaseException;
-import org.sopt.global.security.JwtToken;
-import org.sopt.global.security.JwtTokenProvider;
-import org.sopt.global.security.JwtTokenType;
-import org.sopt.global.security.RefreshTokenHasher;
+import org.sopt.global.security.jwt.JwtToken;
+import org.sopt.global.security.jwt.JwtTokenProvider;
+import org.sopt.global.security.jwt.JwtTokenType;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
