+++ src/main/java/org/sopt/domain/user/presentation/controller/UserCommandControllerDocs.java
@@ -11,7 +11,7 @@ import org.sopt.domain.user.presentation.dto.response.UserResponse;
 import org.sopt.global.annotation.ApiExceptions;
 import org.sopt.global.code.GlobalErrorCode;
 import org.sopt.global.response.CommonApiResponse;
-import org.sopt.global.security.AuthenticatedUser;
+import org.sopt.global.security.authentication.AuthenticatedUser;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.RequestBody;
