+++ src/main/java/org/sopt/global/security/authentication/AuthenticatedUser.java
@@ -1,4 +1,4 @@
-package org.sopt.global.security;
+package org.sopt.global.security.authentication;
 
 /**
  * JWT 인증이 끝난 뒤 컨트롤러에서 사용할 인증 사용자 정보.
