+++ src/main/java/org/sopt/global/security/handler/RestAccessDeniedHandler.java
@@ -1,4 +1,4 @@
-package org.sopt.global.security;
+package org.sopt.global.security.handler;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.http.HttpServletRequest;
@@ -7,33 +7,33 @@ import lombok.RequiredArgsConstructor;
 import org.sopt.global.code.GlobalErrorCode;
 import org.sopt.global.response.CommonApiResponse;
 import org.springframework.http.MediaType;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.AuthenticationEntryPoint;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.web.access.AccessDeniedHandler;
 import org.springframework.stereotype.Component;
 
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 
 /**
- * 인증 실패를 공통 API 응답으로 변환한다.
+ * 인가 실패를 공통 API 응답으로 변환한다.
  */
 @Component
 @RequiredArgsConstructor
-public class RestAuthenticationEntryPoint implements AuthenticationEntryPoint {
+public class RestAccessDeniedHandler implements AccessDeniedHandler {
 
     private final ObjectMapper objectMapper;
 
     @Override
-    public void commence(
+    public void handle(
             HttpServletRequest request,
             HttpServletResponse response,
-            AuthenticationException authException
+            AccessDeniedException accessDeniedException
     ) throws IOException {
-        response.setStatus(GlobalErrorCode.UNAUTHORIZED.getHttpStatus().value());
+        response.setStatus(GlobalErrorCode.FORBIDDEN.getHttpStatus().value());
         response.setCharacterEncoding(StandardCharsets.UTF_8.name());
         response.setContentType(MediaType.APPLICATION_JSON_VALUE);
         response.getWriter().write(objectMapper.writeValueAsString(
-                CommonApiResponse.failureBody(GlobalErrorCode.UNAUTHORIZED)
+                CommonApiResponse.failureBody(GlobalErrorCode.FORBIDDEN)
         ));
     }
 }
