+++ src/main/java/org/sopt/global/security/handler/RestAuthenticationEntryPoint.java
@@ -1,4 +1,4 @@
-package org.sopt.global.security;
+package org.sopt.global.security.handler;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.http.HttpServletRequest;
@@ -7,33 +7,33 @@ import lombok.RequiredArgsConstructor;
 import org.sopt.global.code.GlobalErrorCode;
 import org.sopt.global.response.CommonApiResponse;
 import org.springframework.http.MediaType;
-import org.springframework.security.access.AccessDeniedException;
-import org.springframework.security.web.access.AccessDeniedHandler;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.web.AuthenticationEntryPoint;
 import org.springframework.stereotype.Component;
 
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 
 /**
- * 인가 실패를 공통 API 응답으로 변환한다.
+ * 인증 실패를 공통 API 응답으로 변환한다.
  */
 @Component
 @RequiredArgsConstructor
-public class RestAccessDeniedHandler implements AccessDeniedHandler {
+public class RestAuthenticationEntryPoint implements AuthenticationEntryPoint {
 
     private final ObjectMapper objectMapper;
 
     @Override
-    public void handle(
+    public void commence(
             HttpServletRequest request,
             HttpServletResponse response,
-            AccessDeniedException accessDeniedException
+            AuthenticationException authException
     ) throws IOException {
-        response.setStatus(GlobalErrorCode.FORBIDDEN.getHttpStatus().value());
+        response.setStatus(GlobalErrorCode.UNAUTHORIZED.getHttpStatus().value());
         response.setCharacterEncoding(StandardCharsets.UTF_8.name());
         response.setContentType(MediaType.APPLICATION_JSON_VALUE);
         response.getWriter().write(objectMapper.writeValueAsString(
-                CommonApiResponse.failureBody(GlobalErrorCode.FORBIDDEN)
+                CommonApiResponse.failureBody(GlobalErrorCode.UNAUTHORIZED)
         ));
     }
 }
