+++ src/main/java/org/sopt/global/security/jwt/JwtTokenProvider.java
@@ -1,10 +1,11 @@
-package org.sopt.global.security;
+package org.sopt.global.security.jwt;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.JwtException;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.security.Keys;
 import io.jsonwebtoken.security.WeakKeyException;
+import org.sopt.global.security.exception.JwtAuthenticationException;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import org.springframework.util.StringUtils;
